reafactor: use __SWR_DEVTOOLS_REACT__ to share React with the applications
